refactor(grpc): rename port into bind_port

+++ src/grpcd/opts.rs
@@ -9,7 +9,7 @@ pub struct Opts {
 
     /// Port number that the grpc server is accepting connections on
     #[clap(long)]
-    pub grpc_port: u64,
+    pub grpc_port: u16,
 }
 
 impl Opts {
